Add multi-sensor support to firmware.
This also adds a few "C++isms" to the firmware and will eventually
result in a full C++ification.

Anyway, for now the behavior is that if either of the two sensors
are triggered, the panel is considered pressed and both sensors
will do checks as if they are held until both are released, which
causes the panel to be released.

Some of the refactoring while trying to debug the behavior caused
me to fix the consistency of the naming of the pad (whole device),
panels (arrows you step on), and sensors (electronic devices that
sit under the panels.) An official naming style guide will have to
come at some point.